backend/requirements.txt
opencv-python
numpy
Pillow
scikit-image
tqdm
